+++ libs/ui/src/lib/mocks/holdings.ts
@@ -4,11 +4,11 @@ export const holdings: PortfolioPosition[] = [
   {
     activitiesCount: 1,
     allocationInPercentage: 0.042990776363386086,
-    assetClass: 'EQUITY',
-    assetClassLabel: 'Equity',
     assetProfile: {
       assetClass: 'EQUITY',
+      assetClassLabel: 'Equity',
       assetSubClass: 'STOCK',
+      assetSubClassLabel: 'Stock',
       countries: [
         {
           code: 'US',
@@ -20,60 +20,40 @@ export const holdings: PortfolioPosition[] = [
       currency: 'USD',
       dataSource: 'YAHOO',
       holdings: [],
+      name: 'Apple Inc',
       sectors: [
         {
           name: 'Technology',
           weight: 1
         }
       ],
-      symbol: 'AAPL'
+      symbol: 'AAPL',
+      url: 'https://www.apple.com'
     },
-    assetSubClass: 'STOCK',
-    assetSubClassLabel: 'Stock',
-    countries: [
-      {
-        code: 'US',
-        continent: 'North America',
-        name: 'United States',
-        weight: 1
-      }
-    ],
-    currency: 'USD',
-    dataSource: 'YAHOO',
     dateOfFirstActivity: new Date('2021-12-01T00:00:00.000Z'),
     dividend: 0,
     grossPerformance: 3856,
     grossPerformancePercent: 0.46047289228564603,
     grossPerformancePercentWithCurrencyEffect: 0.46047289228564603,
     grossPerformanceWithCurrencyEffect: 3856,
-    holdings: [],
     investment: 8374,
     marketPrice: 244.6,
-    name: 'Apple Inc',
     netPerformance: 3855,
     netPerformancePercent: 0.460353475041796,
     netPerformancePercentWithCurrencyEffect: 0.036440677966101696,
     netPerformanceWithCurrencyEffect: 430,
     quantity: 50,
-    sectors: [
-      {
-        name: 'Technology',
-        weight: 1
-      }
-    ],
-    symbol: 'AAPL',
     tags: [],
-    url: 'https://www.apple.com',
     valueInBaseCurrency: 12230
   },
   {
     activitiesCount: 2,
     allocationInPercentage: 0.02377401948293552,
-    assetClass: 'EQUITY',
-    assetClassLabel: 'Equity',
     assetProfile: {
       assetClass: 'EQUITY',
+      assetClassLabel: 'Equity',
       assetSubClass: 'STOCK',
+      assetSubClassLabel: 'Stock',
       countries: [
         {
           code: 'DE',
@@ -85,60 +65,40 @@ export const holdings: PortfolioPosition[] = [
       currency: 'EUR',
       dataSource: 'YAHOO',
       holdings: [],
+      name: 'Allianz SE',
       sectors: [
         {
           name: 'Financial Services',
           weight: 1
         }
       ],
-      symbol: 'ALV.DE'
+      symbol: 'ALV.DE',
+      url: 'https://www.allianz.com'
     },
-    assetSubClass: 'STOCK',
-    assetSubClassLabel: 'Stock',
-    countries: [
-      {
-        code: 'DE',
-        continent: 'Europe',
-        name: 'Germany',
-        weight: 1
-      }
-    ],
-    currency: 'EUR',
-    dataSource: 'YAHOO',
     dateOfFirstActivity: new Date('2021-04-23T00:00:00.000Z'),
     dividend: 192,
     grossPerformance: 2226.700251889169,
     grossPerformancePercent: 0.49083842309827874,
     grossPerformancePercentWithCurrencyEffect: 0.29306136948826367,
     grossPerformanceWithCurrencyEffect: 1532.8272791336772,
-    holdings: [],
     investment: 4536.523929471033,
     marketPrice: 322.2,
-    name: 'Allianz SE',
     netPerformance: 2222.2921914357685,
     netPerformancePercent: 0.48986674069961134,
     netPerformancePercentWithCurrencyEffect: 0.034489367670592026,
     netPerformanceWithCurrencyEffect: 225.48257403052068,
     quantity: 20,
-    sectors: [
-      {
-        name: 'Financial Services',
-        weight: 1
-      }
-    ],
-    symbol: 'ALV.DE',
     tags: [],
-    url: 'https://www.allianz.com',
     valueInBaseCurrency: 6763.224181360202
   },
   {
     activitiesCount: 1,
     allocationInPercentage: 0.08038536990007467,
-    assetClass: 'EQUITY',
-    assetClassLabel: 'Equity',
     assetProfile: {
       assetClass: 'EQUITY',
+      assetClassLabel: 'Equity',
       assetSubClass: 'STOCK',
+      assetSubClassLabel: 'Stock',
       countries: [
         {
           code: 'US',
@@ -150,101 +110,73 @@ export const holdings: PortfolioPosition[] = [
       currency: 'USD',
       dataSource: 'YAHOO',
       holdings: [],
+      name: 'Amazon.com, Inc.',
       sectors: [
         {
           name: 'Consumer Discretionary',
           weight: 1
         }
       ],
-      symbol: 'AMZN'
+      symbol: 'AMZN',
+      url: 'https://www.aboutamazon.com'
     },
-    assetSubClass: 'STOCK',
-    assetSubClassLabel: 'Stock',
-    countries: [
-      {
-        code: 'US',
-        continent: 'North America',
-        name: 'United States',
-        weight: 1
-      }
-    ],
-    currency: 'USD',
-    dataSource: 'YAHOO',
     dateOfFirstActivity: new Date('2018-10-01T00:00:00.000Z'),
     dividend: 0,
     grossPerformance: 12758.05,
     grossPerformancePercent: 1.2619300787837724,
     grossPerformancePercentWithCurrencyEffect: 1.2619300787837724,
     grossPerformanceWithCurrencyEffect: 12758.05,
-    holdings: [],
     investment: 10109.95,
     marketPrice: 228.68,
-    name: 'Amazon.com, Inc.',
     netPerformance: 12677.26,
     netPerformancePercent: 1.253938941339967,
     netPerformancePercentWithCurrencyEffect: -0.037866008722316276,
     netPerformanceWithCurrencyEffect: -899.99926757812,
     quantity: 100,
-    sectors: [
-      {
-        name: 'Consumer Discretionary',
-        weight: 1
-      }
-    ],
-    symbol: 'AMZN',
     tags: [],
-    url: 'https://www.aboutamazon.com',
     valueInBaseCurrency: 22868
   },
   {
     activitiesCount: 1,
     allocationInPercentage: 0.19216416482928922,
-    assetClass: 'LIQUIDITY',
-    assetClassLabel: 'Liquidity',
     assetProfile: {
       assetClass: 'LIQUIDITY',
-      assetSubClass: 'CASH',
+      assetClassLabel: 'Liquidity',
+      assetSubClass: 'CRYPTOCURRENCY',
+      assetSubClassLabel: 'Cryptocurrency',
       countries: [],
       currency: 'USD',
       dataSource: 'COINGECKO',
       holdings: [],
+      name: 'Bitcoin',
       sectors: [],
-      symbol: 'bitcoin'
+      symbol: 'bitcoin',
+      url: undefined
     },
-    assetSubClass: 'CRYPTOCURRENCY',
-    assetSubClassLabel: 'Cryptocurrency',
-    countries: [],
-    currency: 'USD',
-    dataSource: 'COINGECKO',
     dateOfFirstActivity: new Date('2017-08-16T00:00:00.000Z'),
     dividend: 0,
     grossPerformance: 52666.7898248,
     grossPerformancePercent: 26.333394912400003,
     grossPerformancePercentWithCurrencyEffect: 26.333394912400003,
     grossPerformanceWithCurrencyEffect: 52666.7898248,
-    holdings: [],
     investment: 1999.9999999999998,
     marketPrice: 97364,
-    name: 'Bitcoin',
     netPerformance: 52636.8898248,
     netPerformancePercent: 26.3184449124,
     netPerformancePercentWithCurrencyEffect: -0.04760906442310894,
     netPerformanceWithCurrencyEffect: -2732.737808972287,
     quantity: 0.5614682,
-    sectors: [],
-    symbol: 'bitcoin',
     tags: [],
-    url: undefined,
     valueInBaseCurrency: 54666.7898248
   },
   {
     activitiesCount: 1,
     allocationInPercentage: 0.04307127421937313,
-    assetClass: 'EQUITY',
-    assetClassLabel: 'Equity',
     assetProfile: {
       assetClass: 'EQUITY',
+      assetClassLabel: 'Equity',
       assetSubClass: 'STOCK',
+      assetSubClassLabel: 'Stock',
       countries: [
         {
           code: 'US',
@@ -256,60 +188,40 @@ export const holdings: PortfolioPosition[] = [
       currency: 'USD',
       dataSource: 'YAHOO',
       holdings: [],
+      name: 'Microsoft Corporation',
       sectors: [
         {
           name: 'Technology',
           weight: 1
         }
       ],
-      symbol: 'MSFT'
+      symbol: 'MSFT',
+      url: 'https://www.microsoft.com'
     },
-    assetSubClass: 'STOCK',
-    assetSubClassLabel: 'Stock',
-    countries: [
-      {
-        code: 'US',
-        continent: 'North America',
-        name: 'United States',
-        weight: 1
-      }
-    ],
-    currency: 'USD',
-    dataSource: 'YAHOO',
     dateOfFirstActivity: new Date('2023-01-03T00:00:00.000Z'),
     dividend: 0,
     grossPerformance: 5065.5,
     grossPerformancePercent: 0.7047750229568411,
     grossPerformancePercentWithCurrencyEffect: 0.7047750229568411,
     grossPerformanceWithCurrencyEffect: 5065.5,
-    holdings: [],
     investment: 7187.4,
     marketPrice: 408.43,
-    name: 'Microsoft Corporation',
     netPerformance: 5065.5,
     netPerformancePercent: 0.7047750229568411,
     netPerformancePercentWithCurrencyEffect: -0.015973588391056275,
     netPerformanceWithCurrencyEffect: -198.899926757814,
     quantity: 30,
-    sectors: [
-      {
-        name: 'Technology',
-        weight: 1
-      }
-    ],
-    symbol: 'MSFT',
     tags: [],
-    url: 'https://www.microsoft.com',
     valueInBaseCurrency: 12252.9
   },
   {
     activitiesCount: 1,
     allocationInPercentage: 0.18762679306394897,
-    assetClass: 'EQUITY',
-    assetClassLabel: 'Equity',
     assetProfile: {
       assetClass: 'EQUITY',
+      assetClassLabel: 'Equity',
       assetSubClass: 'STOCK',
+      assetSubClassLabel: 'Stock',
       countries: [
         {
           code: 'US',
@@ -321,60 +233,40 @@ export const holdings: PortfolioPosition[] = [
       currency: 'USD',
       dataSource: 'YAHOO',
       holdings: [],
+      name: 'Tesla, Inc.',
       sectors: [
         {
           name: 'Consumer Discretionary',
           weight: 1
         }
       ],
-      symbol: 'TSLA'
+      symbol: 'TSLA',
+      url: 'https://www.tesla.com'
     },
-    assetSubClass: 'STOCK',
-    assetSubClassLabel: 'Stock',
-    countries: [
-      {
-        code: 'US',
-        continent: 'North America',
-        name: 'United States',
-        weight: 1
-      }
-    ],
-    currency: 'USD',
-    dataSource: 'YAHOO',
     dateOfFirstActivity: new Date('2017-01-03T00:00:00.000Z'),
     dividend: 0,
     grossPerformance: 51227.500000005,
     grossPerformancePercent: 23.843379101756675,
     grossPerformancePercentWithCurrencyEffect: 23.843379101756675,
     grossPerformanceWithCurrencyEffect: 51227.500000005,
-    holdings: [],
     investment: 2148.499999995,
     marketPrice: 355.84,
-    name: 'Tesla, Inc.',
     netPerformance: 51197.500000005,
     netPerformancePercent: 23.829415871596066,
     netPerformancePercentWithCurrencyEffect: -0.12051410125545206,
     netPerformanceWithCurrencyEffect: -7314.00091552734,
     quantity: 150,
-    sectors: [
-      {
-        name: 'Consumer Discretionary',
-        weight: 1
-      }
-    ],
-    symbol: 'TSLA',
     tags: [],
-    url: 'https://www.tesla.com',
     valueInBaseCurrency: 53376
   },
   {
     activitiesCount: 5,
     allocationInPercentage: 0.053051250766657634,
-    assetClass: 'EQUITY',
-    assetClassLabel: 'Equity',
     assetProfile: {
       assetClass: 'EQUITY',
+      assetClassLabel: 'Equity',
       assetSubClass: 'ETF',
+      assetSubClassLabel: 'ETF',
       countries: [
         {
           code: 'US',
@@ -386,50 +278,30 @@ export const holdings: PortfolioPosition[] = [
       currency: 'USD',
       dataSource: 'YAHOO',
       holdings: [],
+      name: 'Vanguard Total Stock Market Index Fund ETF Shares',
       sectors: [
         {
           name: 'Equity',
           weight: 1
         }
       ],
-      symbol: 'VTI'
+      symbol: 'VTI',
+      url: 'https://www.vanguard.com'
     },
-    assetSubClass: 'ETF',
-    assetSubClassLabel: 'ETF',
-    countries: [
-      {
-        code: 'US',
-        weight: 1,
-        continent: 'North America',
-        name: 'United States'
-      }
-    ],
-    currency: 'USD',
-    dataSource: 'YAHOO',
     dateOfFirstActivity: new Date('2019-03-01T00:00:00.000Z'),
     dividend: 0,
     grossPerformance: 6845.8,
     grossPerformancePercent: 1.0164758094605268,
     grossPerformancePercentWithCurrencyEffect: 1.0164758094605268,
     grossPerformanceWithCurrencyEffect: 6845.8,
-    holdings: [],
     investment: 8246.2,
     marketPrice: 301.84,
-    name: 'Vanguard Total Stock Market Index Fund ETF Shares',
     netPerformance: 6746.3,
     netPerformancePercent: 1.0017018833976383,
     netPerformancePercentWithCurrencyEffect: 0.01085061564051406,
     netPerformanceWithCurrencyEffect: 161.99969482422,
     quantity: 50,
-    sectors: [
-      {
-        name: 'Equity',
-        weight: 1
-      }
-    ],
-    symbol: 'VTI',
     tags: [],
-    url: 'https://www.vanguard.com',
     valueInBaseCurrency: 15092
   }
 ];